+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * `TabModel` now supports a bindable `tooltip`, which can be used to render strings or elements
   while hovering over tabs.
 * New `Placeholder` component provides a thin wrapper around `Box` with standardized, muted styling.
-* New `matchMode` prop on `StoreFilterField` to allow customizing match to: `start`, `startWord`, or `any`.
+* New `StoreFilterField.matchMode` prop allows customizing match to `start`, `startWord`, or `any`.
 * `Select` now implements enhanced typeahead filtering of options. The default filtering is now
   based on a case-insensitive match of word starts in the label. (Previously it was based on a match
   _anywhere_ in the label _or_ value.) To customize this behavior, applications should use the new